recommandation-service/requirements.txt
django
djangorestframework
djangorestframework-simplejwt

python-consul
requests
mysqlclient